Major refactor of GenerateGOAnnotationFile to make it more testable
